# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing passengers and taxists in a ride-sharing application.",
        "title": "Ride-Sharing Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.55.42:8080",
    "basePath": "/",
    "paths": {
...
        "/protected/announcements": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "add a ugur",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "POST a new Ugur for taxist",
                "parameters": [
                    {
                        "description": "Ugur",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TaxistAnnouncmentRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/protected/reserve-packages/{taxi_ann_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "add a new reserving package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "POST a new reserving package for passenger user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ugur ID",
                        "name": "taxi_ann_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reserve Packages",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReservePackages"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/protected/reserve-passengers/{taxi_ann_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "add a new reserving passenger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "POST a new reserving passenger for passenger user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ugur ID",
                        "name": "taxi_ann_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reserve Passenger",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PostReservePassengers"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/protected/ugurlar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of all ugurlar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Get all ugurlar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/protected/ugurlar/{ugur_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve ugur by given ugur_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Get ugur by ugur_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ugur ID",
                        "name": "ugur_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "handlers.PassengerPeople": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "handlers.PostReservePassengers": {
            "type": "object",
            "required": [
                "package",
                "passengers"
            ],
            "properties": {
                "package": {
                    "type": "string"
                },
                "passengers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.PassengerPeople"
                    }
                }
            }
        },
...
        "models.AnnouncementType": {
            "type": "string",
            "enum": [
                "person",
                "package",
                "person_and_package"
            ],
            "x-enum-varnames": [
                "Person",
                "Package",
                "PersonAndPackage"
            ]
        },
...
        "models.ReservePackages": {
            "type": "object",
            "properties": {
                "about_package": {
                    "type": "string"
                },
                "package_reciever": {
                    "type": "string"
                },
                "package_sender": {
                    "type": "string"
                },
                "reciever_phone": {
                    "type": "string"
                },
                "sender_phone": {
                    "type": "string"
                }
            }
        },
        "models.TaxistAnnouncmentRequest": {
            "type": "object",
            "required": [
                "depart_date",
                "depart_time",
                "from_place",
                "space",
                "to_place",
                "type"
            ],
            "properties": {
                "depart_date": {
                    "type": "string",
                    "example": "2025-01-01"
                },
                "depart_time": {
                    "type": "string",
                    "example": "15:30:00"
                },
                "from_place": {
                    "type": "integer",
                    "example": 0
                },
                "space": {
                    "type": "integer",
                    "example": 4
                },
                "to_place": {
                    "type": "integer",
                    "example": 0
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.AnnouncementType"
                        }
                    ],
                    "example": "submit one of person, package or person_and_package"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}